% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmics.R
\name{coef.multiOmics}
\alias{coef.multiOmics}
\title{coef.multiOmics}
\usage{
\method{coef}{multiOmics}(multiOmics_mod)
}
\arguments{
\item{multiOmics_mod}{}
}
\description{
coef method for multiOmics model object to obtain clinical feature coefficients
}
