% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_stage.R
\name{.folds}
\alias{.folds}
\title{create folds}
\usage{
.folds(x, num_fold)
}
\arguments{
\item{x}{a vector of integers to split}

\item{num_fold}{number of folds. integer k}
}
\description{
create folds
}
