% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eg_gene_symbols}
\alias{eg_gene_symbols}
\title{Example Gene Names}
\format{
An object of class \code{character} of length 1000.
}
\usage{
eg_gene_symbols
}
\description{
use with \code{get_grouping}
}
\keyword{datasets}
