% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eg_multiOmics}
\alias{eg_multiOmics}
\title{Example multiOmics result}
\format{
An object of class \code{multiOmics} of length 4.
}
\usage{
eg_multiOmics
}
\description{
Example \code{multiOmics} and \code{multiOmics_sensitivity} from vignette("EMMultiOmics")
\itemize{
\item eg_multiOmics
\item eg_multiOmics_sensitivity
}
}
\keyword{datasets}
