% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMVS.R
\name{EMVS}
\alias{EMVS}
\title{Expectation Maximization Variable Selection Function}
\usage{
EMVS(
  M,
  G,
  grouping,
  nu0 = 0.5,
  nu1 = 10^3,
  nu = 1,
  lambda = 1,
  a = 1,
  b = 1,
  I = 100,
  THRESH = 1e-04
)
}
\arguments{
\item{M}{DNA Methylation matrix}

\item{G}{Gene expression level}

\item{grouping}{Gene grouping}

\item{nu0}{...}

\item{nu1}{...}

\item{lambda}{...}

\item{a}{...}

\item{b}{...}

\item{I}{number of iterations...}

\item{THRESH}{threshold for ...}
}
\description{
Performs EMVS algorithm for genomics data.
}
\details{
\code{EMVS_par()} uses parallel backend for computation.
}
\examples{

# example data
G <- GBM_data2$G
C <- GBM_data2$C
Y <- log(GBM_data2$Y)
M <- GBM_data2$M
DELTA <- GBM_data2$Delta
grouping <- GENE_GROUP2

EMVS(M, G, grouping, I=10, THRESH=0.001)
EMVS_par(M, G, grouping, I=10, THRESH=0.001)

}
