% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GBM2_full_EMVS_res}
\alias{GBM2_full_EMVS_res}
\title{GBM2_full_EMVS_res}
\format{
GBM 2 full data EMVS result
}
\usage{
GBM2_full_EMVS_res
}
\description{
GBM2_full_EMVS_res
}
\keyword{datasets}
