% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GBM_data2}
\alias{GBM_data2}
\title{GBM_data2}
\format{
An object of class \code{list} of length 5.
}
\usage{
GBM_data2
}
\description{
Subset of GBM 2 data

includes the following objects:
\itemize{
\item GBM_data2: methylation (M), gene expression (G), clinical features (C), and clinical outcomes (Y)
\item GENE_GROUP2: gene grouping information
\item GBM2_EMVS_res: EVMS result
}
}
\keyword{datasets}
