% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GBM_data2_full}
\alias{GBM_data2_full}
\title{GBM_data2_full}
\format{
An object of class \code{list} of length 5.
}
\usage{
GBM_data2_full
}
\description{
Full GBM 2 data
\itemize{
\item GBM_data2_full
\item GENE_GROUP2_FULL
\item GBM2_full_EMVS_res
}
}
\keyword{datasets}
