% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GENE_GROUP2_FULL}
\alias{GENE_GROUP2_FULL}
\title{"GENE_GROUP2_FULL"}
\format{
Gene grouping for GBM full data 2
}
\usage{
GENE_GROUP2_FULL
}
\description{
"GENE_GROUP2_FULL"
}
\keyword{datasets}
