% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grouping.R
\name{get_grouping}
\alias{get_grouping}
\title{Get Gene Groupings from Functional classification text file}
\usage{
get_grouping(full_gene_list, file_dir)
}
\description{
Get Gene Groupings from Functional classification text file
}
