% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_stage.R
\name{lpcf}
\alias{lpcf}
\title{LPCF LOG Parabolic Cylinder Function}
\usage{
lpcf(k, v, z, .mpmath)
}
\arguments{
\item{k}{details...}

\item{v}{details...}

\item{z}{details...}

\item{.mpmath}{}
}
\description{
The function has dependency on python pacakge: mpmath
}
