% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmics.R
\name{multiOmics_sensitivity}
\alias{multiOmics_sensitivity}
\title{multiOmics_sensitivity}
\usage{
multiOmics_sensitivity(
  gstr_vec,
  a0_vec,
  M,
  G,
  grouping,
  nu0 = 0.5,
  nu1 = 10^3,
  nu = 1,
  lambda = 1,
  a = 1,
  b = 1,
  EMVS_I = 10,
  EMVS_thresh = 1e-04,
  Y,
  C,
  Delta,
  n_fold = 10,
  random_seed = NULL,
  NEG_I = 10,
  NEG_thresh = 1e-04
)
}
\arguments{
\item{gstr_vec}{a vector of \code{g} values (if gstr == "scale", then gstr = 1/N^2, where N=nrow(G))}

\item{a0_vec}{a vector of \code{a} values}

\item{...}{Additional parameters for \code{multiOmics}params

M <- GBM_data2$M
G <- GBM_data2$G
fp <- system.file("eg_fx_classification.txt", package="EMMultiOmics")
gene_grouping <- get_grouping(eg_gene_symbols, fp)
Y <- GBM_data2$Y
C <- GBM_data2$C
Delta <- GBM_data2$Delta
n_fold <- 10
random_seed <- 123
EMVS_I <- 3
NEG_I <- 3
EMVS_thresh <- 0.0001
NEG_thresh <- 0.0001

a0 <- c(0.1, 1, 10, 50)
g <- list("scale", 1)run

multiOmics_s_obj <- multiOmics_sensitivity(
M=M,
G=G,
grouping=grouping,
Y=Y,
C=C,
a0_vec=a0,
gstr_vec=g,
Delta=Delta,
n_fold=n_fold,
random_seed = random_seed,
EMVS_I = EMVS_I,
NEG_I=NEG_I,
EMVS_thresh = EMVS_thresh,
NEG_thresh = NEG_thresh
)summary

summary(multiOmics_s_obj)}
}
\description{
Run \code{multiOmics} for given sets of parameters \code{g} and \code{a}
}
