% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_stage.R
\name{NEG_em}
\alias{NEG_em}
\title{Normal-Exponential-Gamma EM Algorithm}
\usage{
NEG_em(
  Y,
  G,
  C,
  a0,
  gstr,
  Zmatrix,
  I = 10,
  thresh = 0.001,
  .mpmath = setup_mpmath()
)
}
\arguments{
\item{Y}{clinical outcome}

\item{G}{gene expression level}

\item{C}{clinical features}

\item{a0}{size sparsity}

\item{gstr}{prior (if "scale", then gstr = 1/N^2, where N=nrow(G))}

\item{Zmatrix}{Loading matrix}

\item{I}{number of maximum iterations}

\item{thresh}{convergence criterion}

\item{.mpmath}{function depends on mpmath package from python. pointer for mpmath package}
}
\description{
A short description...
}
\examples{
G <- GBM_data2$G
Y <- GBM_data2$Y
C <- GBM_data2$C
R2 <- GBM2_EMVS_res$R2
Zmatrix <- Zmat_builder(R2, G)
a0 <- 0.1
gstr <- "scale"
mpmath <- setup_mpmath()

NEG_em(Y, G, C, a0, gstr, Zmatrix,
       I=10, thresh=0.001, .mpmath=mpmath)
}
