% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmics.R
\name{plot.multiOmics}
\alias{plot.multiOmics}
\title{plot.multiOmics}
\usage{
\method{plot}{multiOmics}(multiOmics_mod)
}
\arguments{
\item{multiOmics_mod}{multiOmics model object}
}
\description{
Summary box plot of multiOmics model object
}
