% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmics.R
\name{plot.multiOmics_sensitivity}
\alias{plot.multiOmics_sensitivity}
\title{plot.multiOmics_sensitivity}
\usage{
\method{plot}{multiOmics_sensitivity}(multiOmics_sensitivity_obj)
}
\arguments{
\item{multiOmics_sensitivity_obj}{multiOmics_sensitivity object}
}
\description{
plot function for multiOmics_sensitivity object
}
