% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmics.R
\name{print.multiOmics}
\alias{print.multiOmics}
\title{print.multiOmics}
\usage{
\method{print}{multiOmics}(multiOmics_mod)
}
\arguments{
\item{multiOmics_mod}{multiOmics model object}
}
\description{
print method for multiOmics model object
}
