% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_stage.R
\name{second_stage_helper}
\alias{second_stage_helper}
\title{2nd stage modeling helper}
\usage{
second_stage_helper(
  Y,
  G,
  C,
  Delta,
  R2,
  a0,
  gstr,
  n_fold = 10,
  random_seed = NULL,
  ...
)
}
\arguments{
\item{Y}{details..}

\item{G}{details..}

\item{C}{details..}

\item{Delta}{details...}

\item{R2}{vector of R2 (R-squared?) from first EMVS algorithm}

\item{a0}{details... (or this can be under \code{...})}

\item{gstr}{details... (or this can be under \code{...})}

\item{n_fold}{number of folds in K-fold cross validation}

\item{random_seed}{random seed for splitting folds for cross validation}

\item{...}{to supply any additional arguments to NEG_em, the arguments are \code{I} and \code{thresh}}
}
\description{
Runs cross validation for 2nd stage modeling.
}
\details{
... can take:
\itemize{
\item \code{I} = number of iterations for \code{NEG_em}
\item \code{thresh} = convergence criterion for \code{NEG_em}
}
}
\examples{
G <- GBM_data2$G
Y <- GBM_data2$Y
C <- GBM_data2$C
Delta <- GBM_data2$Delta
R2 <- GBM2_EMVS_res$R2
a0 <- 0.1
gstr <- "scale"
n_fold <- 10
random_seed <- 123

second_stage_helper(G, Y, C, Delta, R2, a0, gstr, n_fold, random_seed,
                    I=10, thresh=0.001)

# multiple a0
a0 <- c(0.1, 1, 10, 50)
lapply(a0, function(a) second_stage_helper(Y, G, C, Delta, R2, a0=a, gstr, n_fold, random_seed, 
                                           I=10, thresh=0.001)) |> 
  setNames(a0)

}
