% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_stage.R
\name{Zmat_builder}
\alias{Zmat_builder}
\title{Z-matrix builder}
\usage{
Zmat_builder(R2, G)
}
\arguments{
\item{R2}{R-squared from EMVS result}

\item{G}{gene expression level}
}
\description{
Builds Z matrix to be used with NEG_em
}
\examples{
G <- GBM_data2$G
R2 <- GBM2_EMVS_res$R2

Zmat_builder(R2, G)

}
