% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eg_gene_symbols}
\alias{eg_gene_symbols}
\title{Example Gene Names}
\format{
Example Gene ID's from Official Symbols
}
\usage{
eg_gene_symbols
}
\description{
Example Gene Names
}
\keyword{datasets}
