% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMVS.R
\name{EMVS}
\alias{EMVS}
\title{Expectation Maximization Variable Selection Function}
\usage{
EMVS(
  M,
  G,
  grouping,
  nu0 = 0.5,
  nu1 = 10^3,
  nu = 1,
  lambda = 1,
  a = 1,
  b = 1,
  I = 100,
  THRESH = 1e-04
)
}
\arguments{
\item{M}{DNA Methylation matrix}

\item{G}{Gene expression level}

\item{grouping}{Gene grouping}

\item{nu0}{parameter 0 for spike-and-slab Gaussian mixture prior on \eqn{\beta}}

\item{nu1}{parameter 1 for spike-and-slab Gaussian mixture prior on \eqn{\beta}}

\item{lambda}{For the prior on \eqn{\sigma2}, an inverse gamma prior \eqn{\pi(\sigma2 | \gamma) = IG(\nu/2, \nu\lambda/2)}, <- what is lambda here?}

\item{a}{Parameter \eqn{a} for the beta prior \eqn{\pi(\theta) \propto \theta a-1(1-\theta)b-1}, \eqn{a, b > 0}}

\item{b}{Parameter \eqn{b} for the beta prior \eqn{\pi(\theta) \propto \theta a-1(1-\theta)b-1}, \eqn{a, b > 0}}

\item{I}{Maximum number of iterations of EMVS}

\item{THRESH}{Threshold for change of parameters. The threshold will determine convergence.}
}
\description{
Performs EMVS algorithm for genomics data.
The default values for \code{nu0}, \code{nu1}, \code{lambda}, \code{a}, \code{b} were chosen per publication: \emph{Veronika Ročková & Edward I. George (2013): EMVS: The EM Approach to Bayesian Variable Selection, Journal of the American Statistical Association}
}
\examples{

# example data
M <- GBM_data2$M
G <- GBM_data2$G
grouping <- GENE_GROUP2

EMVS(M, G, grouping, I=10, THRESH=0.001)
EMVS_par(M, G, grouping, I=10, THRESH=0.001)

}
