% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grouping.R
\name{get_grouping}
\alias{get_grouping}
\title{Get Gene Groupings from Functional classification text file}
\usage{
get_grouping(gene_symbols, file_dir)
}
\arguments{
\item{gene_symbols}{a vector of gene names as official symbols.}

\item{file_dir}{file path of functional classification obtained from DAVID tool. A txt file is expected. Please see vignette("gene_grouping") for guide}
}
\description{
Given a functional classification result of genes, the function divides genes into groups
}
\examples{
fp <- system.file("eg_fx_classification.txt", package="EMMultiOmics")
get_grouping(eg_gene_symbols, fp)

}
