% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmics.R
\name{summary.multiOmics_sensitivity}
\alias{summary.multiOmics_sensitivity}
\title{summary.multiOmics_sensitivity}
\usage{
\method{summary}{multiOmics_sensitivity}(multiOmics_sensitivity_obj)
}
\arguments{
\item{multiOmics_sensitivity_obj}{multiOmics_sensitivity object}
}
\description{
summary method for multiOmics_sensitivity object
}
